+++ sync/sync.go
@@ -10,7 +10,9 @@ import (
 
 // Some of the code here is inspired by(or taken from):
 //   (a) https://github.com/golang/sync/tree/v0.3.0/errgroup whose license(BSD 3-Clause) can be found here: https://github.com/golang/sync/blob/v0.3.0/LICENSE
-//   (b) https://github.com/sourcegraph/conc whose license(MIT) can be found here:                          https://github.com/sourcegraph/conc/blob/v0.3.0/LICENSE
+//   (b) https://go-review.googlesource.com/c/sync/+/416555
+//   (c) https://github.com/sourcegraph/conc whose license(MIT) can be found here:                          https://github.com/sourcegraph/conc/blob/v0.3.0/LICENSE
+//   (d) https://github.com/kucherenkovova/safegroup whose license(BSD 3-Clause) can be found here:         https://github.com/kucherenkovova/safegroup/blob/v1.0.2/LICENSE
 //
 
 // Group is a datastructure that waits for a collection of goroutines to finish.
@@ -30,13 +32,14 @@ type Group struct {
 	errMu         sync.Mutex // protects err
 	err           error
 	collectedErrs []error
+	panic         interface{} // PanicError or PanicValue
 }
 
 // New returns a valid Group and a context(derived from ctx).
 // The Group limits the number of active goroutines to at most n.
 //
 // The derived Context is canceled the first time Go returns.
-// Unlike [golang.org/x/sync/errgroup.Group] it is not cancelled the first time a function passed to Go returns an error.
+// Unlike [golang.org/x/sync/errgroup.Group] it is not cancelled the first time a function passed to Go returns an error or panics.
 //
 // n limits the number of active goroutines in this Group.
 // If n is <=0, it indicates no limit.
@@ -55,7 +58,8 @@ func New(ctx context.Context, n int) (*Group, context.Context) {
 // active goroutines in the Group exceeding the configured limit.
 //
 // It also blocks until all function calls from the Go method have returned, then returns the concated non-nil errors(if any) from them.
-// Unlike [golang.org/x/sync/errgroup.Group] the first call to return an error does not cancel the Group's context.
+// If any of those functions panic, Go will also propagate that panic.
+// Unlike [golang.org/x/sync/errgroup.Group] the first call to return an error(or panics) does not cancel the Group's context.
 //
 // If called concurrently, it will block until the previous call returns.
 func (w *Group) Go(funcs ...func() error) error {
@@ -64,6 +68,10 @@ func (w *Group) Go(funcs ...func() error) error {
 		if w.cancel != nil {
 			w.cancel(w.err)
 		}
+		if w.panic != nil {
+			panic(w.panic)
+		}
+
 		return nil
 	}
 
@@ -91,6 +99,10 @@ func (w *Group) Go(funcs ...func() error) error {
 			if w.cancel != nil {
 				w.cancel(w.err)
 			}
+			if w.panic != nil {
+				panic(w.panic)
+			}
+
 			return w.err
 		}
 	}
@@ -132,11 +144,19 @@ func (w *Group) Go(funcs ...func() error) error {
 		if w.cancel != nil {
 			w.cancel(w.err)
 		}
+		if w.panic != nil {
+			panic(w.panic)
+		}
+
 		return w.err
 	}
 }
 
 func (w *Group) done() {
+	if v := recover(); v != nil {
+		w.panic = addStack(v)
+	}
+
 	if w.sem != nil {
 		<-w.sem
 	}